+++ src/components/App/Footer/index.js
